==> tactical_analyses/00.2_annotation-distances.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Setup 

```{r}

"%&%" <- function(a,b) paste0(a,b)

library("tidyverse")
library("data.table")
library("GenomicRanges")

serv0.dir <- "/home/jason/science/servers/FUSE/"
serv.dir <- "/home/jason/science/servers/FUSE5/"
proj.dir <- serv.dir %&% "projects/t2d_classification/"
write.dir <- proj.dir %&% "tactical_analyses/"
plot.dir <- write.dir %&% "plots/"


bed.dir <- serv0.dir %&% "reference/chromatin_segmentation/varshney_2017/chromatin_states/" 


```


```{r}

Active_enhancer_1 <- c("9_Active_enhancer_1")
Active_enhancer_2 <- c("10_Active_enhancer_2")
Weak_enhancer <- c("11_Weak_enhancer")
repressed.polycomb <- c("16_Repressed_polycomb")
repressed.polycomb.weak <- c("17_Weak_repressed_polycomb")
Active_TSS <- c("1_Active_TSS")
Weak_TSS <- c("2_Weak_TSS")
Flanking_TSS <- c("3_Flanking_TSS")
Strong_transcription <- c("5_Strong_transcription")
Weak_transcription <- c("6_Weak_transcription")
Bivalent_poised_TSS <- c("14_Bivalent_poised_TSS")
Genic_enhancer <- c("8_Genic_enhancer")
low.signal <- c("18_Quiescent_low_signal")
Coding <- c("coding")

```

```{r}

tiss.vec <- c("Islets","Adipose","SkeletalMuscle","Liver")
isl.df <- fread("cat " %&% bed.dir %&% "Islets.chromatinStates.bed.gz | zmore",sep="\t")
adi.df <- fread("cat " %&% bed.dir %&% "Adipose.chromatinStates.bed.gz | zmore",sep="\t")
mus.df <- fread("cat " %&% bed.dir %&% "SkeletalMuscle.chromatinStates.bed.gz | zmore",sep="\t")
liv.df <- fread("cat " %&% bed.dir %&% "Liver.chromatinStates.bed.gz | zmore",sep="\t")

isl.df$V4 <- isl.df$V4 %>% gsub("/","_",.)
adi.df$V4 <- adi.df$V4 %>% gsub("/","_",.)
mus.df$V4 <- mus.df$V4 %>% gsub("/","_",.)
liv.df$V4 <- liv.df$V4 %>% gsub("/","_",.)

```



```{r}

build_annotation_count_df <- function(tiss.df){
  df.list <- list(isl.df,adi.df,mus.df,liv.df)
  tiss.vec <- c("Islet","Adipose","Muscle","Liver")
  annot.list <- list(Active_enhancer_1,Active_enhancer_2,Weak_enhancer,
                    Active_TSS,Weak_TSS, Flanking_TSS,
                    Strong_transcription,Weak_transcription,
                    Genic_enhancer)
  name.vec <- c("Active_enhancer_1","Active_enhancer_2","Weak_enhancer",
                 "Active_TSS","Weak_TSS", "Flanking_TSS",
                 "Strong_transcription","Weak_transcription",
                 "Genic_enhancer")
  out.df <- c()
  pb <- txtProgressBar(min=0,max=length(df.list),style=3)
  for (i in 1:length(df.list)){
    setTxtProgressBar(pb,i)
    df <- df.list[[i]]
    tot <- dim(df)[1]
    tiss <- tiss.vec[i]
    for (e in 1:length(annot.list)){
      annot <- name.vec[e]
      a.vec <- annot.list[[e]]
      sub <- filter(df,V4 %in% a.vec)
      count <- dim(sub)[1]
      prop <- count/tot
      build.df <- data.frame(Tissue=tiss,Annotation=annot,Count=count,
                             Proportion=prop,stringsAsFactors = F)
      out.df <- rbind(out.df,build.df)
    }
  }
  
  return(out.df)
}

count.df <- build_annotation_count_df()
write.table(x=count.df,file=write.dir %&%
              "analysis_files/annot_count.txt",sep="\t",quote=F,row.names=F)
```




```{r}

convert_to_gr <- function(df){
  gr <- GRanges(seqnames=df$V1,IRanges(start=df$V2,end=df$V3)) %>% reduce(.)
}

get_mean_nearest_distances <- function(gr1,gr2){
  d <- (distanceToNearest(gr1,gr2) %>% as.data.frame(.))$distance %>% mean(.)
  d <- (distanceToNearest(gr2,gr1) %>% as.data.frame(.))$distance %>% mean(.) 
}

get_median_nearest_distances <- function(gr1,gr2){
  d <- (distanceToNearest(gr1,gr2) %>% as.data.frame(.))$distance %>% median(.)
  d <- (distanceToNearest(gr2,gr1) %>% as.data.frame(.))$distance %>% median(.) 
}

```



```{r}


build_dist_df <- function(tiss.df){
  se.gr1 <- filter(tiss.df,V4 %in% Active_enhancer_1) %>% convert_to_gr(.) 
  se.gr2 <- filter(tiss.df,V4 %in% Active_enhancer_2) %>% convert_to_gr(.) 
  we.gr <- filter(tiss.df,V4 %in% Weak_enhancer) %>% convert_to_gr(.) 
  sp.gr <- filter(tiss.df,V4 %in% Active_TSS) %>% convert_to_gr(.) 
  wp.gr <- filter(tiss.df,V4 %in% Weak_TSS) %>% convert_to_gr(.) 
  fp.gr <-  filter(tiss.df,V4 %in% Flanking_TSS) %>% convert_to_gr(.) 
  st.gr <- filter(tiss.df,V4 %in% Strong_transcription) %>% convert_to_gr(.)
  wt.gr <- filter(tiss.df,V4 %in% Weak_transcription) %>% convert_to_gr(.)
  ge.gr <- filter(tiss.df,V4 %in% Genic_enhancer) %>% convert_to_gr(.) 
  
  gr.list <- list(se.gr1,se.gr2,we.gr,sp.gr,wp.gr,fp.gr,st.gr,wt.gr,ge.gr)
  name.vec <- c("Active_enhancer_1","Active_enhancer_2","Weak_enhancer",
                 "Active_TSS","Weak_TSS", "Flanking_TSS",
                 "Strong_transcription","Weak_transcription",
                 "Genic_enhancer")
  pb <- txtProgressBar(min=0,max=length(name.vec),style=3)
  out.df <- c()
  for (i in 1:length(name.vec)){
    setTxtProgressBar(pb,i)
    for (e in 1:length(name.vec)){
      ref.annot <- name.vec[i]
      query.annot <- name.vec[e]
      mean.dist <- get_mean_nearest_distances(gr.list[[i]],gr.list[[e]])
      median.dist <- get_median_nearest_distances(gr.list[[i]],gr.list[[e]])
      build.df <- data.frame(ref.annot,query.annot,mean.dist,median.dist,
                             stringsAsFactors = FALSE)
      out.df <- rbind(out.df,build.df)
    }
  }  
  out.df$mean.dist.K <- (out.df$mean.dist/1000) %>% round(.,digits=1)
  out.df$median.dist.K <- (out.df$median.dist/1000) %>% round(.,digits=1)
  return(out.df)
}


```


```{r}

median_tile_plot <- function(dist.df,omit.vec=c("Genic_enhancer","Flanking_TSS")){
  p <- ggplot(filter(dist.df,!(ref.annot%in%omit.vec),!(query.annot%in%omit.vec)),
               aes(ref.annot, query.annot)) +
    geom_tile(aes(fill = median.dist.K), color = "white") +
    geom_text(aes(label=median.dist.K)) + 
    scale_fill_gradientn(colours=rev(viridis(100)[1:92])) +
    theme(legend.title = element_text(size = 10),
          legend.text = element_text(size = 12),
          plot.title = element_text(size=16),
          axis.title=element_blank(),
          axis.text.x = element_text(angle = 30, hjust = 1)) +
    labs(fill = "Median distance\nto nearest (Kb)")  
  return(p)
}


mean_tile_plot <- function(dist.df,omit.vec=c("Genic_enhancer","Flanking_TSS")){
  p <- ggplot(filter(dist.df,!(ref.annot%in%omit.vec),!(query.annot%in%omit.vec)),
               aes(ref.annot, query.annot)) +
    geom_tile(aes(fill = mean.dist.K), color = "white") +
    geom_text(aes(label=mean.dist.K)) + 
    scale_fill_gradientn(colours=rev(viridis(100)[1:92])) +
    theme(legend.title = element_text(size = 10),
          legend.text = element_text(size = 12),
          plot.title = element_text(size=16),
          axis.title=element_blank(),
          axis.text.x = element_text(angle = 30, hjust = 1)) +
    labs(fill = "Mean distance\nto nearest (Kb)")  
  return(p)
}


```


```{r}

library("grid")
library("gridExtra")
library("viridis")


save_composite_plots <- function(){
  df.list <- list(isl.df,adi.df,mus.df,liv.df)
  tiss.vec <- c("Islet","Adipose","Muscle","Liver")
  for (i in 1:length(tiss.vec)){
    tiss <- tiss.vec[i]
    print(tiss)
    dist.df <- build_dist_df(df.list[[i]])
    write.table(x=dist.df,file=write.dir %&% "analysis_files/annot_dists_" %&% 
                  tiss %&% ".txt",sep="\t",quote=F,row.names=F)
    plot.df <- filter(count.df,Tissue==tiss) %>% arrange(.,desc(Count))
    plot.df$Proportion <- prettyNum(plot.df$Proportion,digits=2)
    plot.df$Percentage <- as.character(as.numeric(plot.df$Proportion) * 100 ) %&% "%"
    plot.df$Annotation <- factor(plot.df$Annotation,levels=plot.df$Annotation)
    p1 <- ggplot(data=plot.df,aes(x=Annotation,y=Count/1000)) +
      geom_bar(stat="identity",color="black",fill=viridis(10)[2]) + 
      geom_text(aes(x=Annotation,y=(Count/1000)+5,label=Percentage)) +  
      theme(axis.text.x = element_text(angle=30,vjust=0.7)) + 
      ylab("Count (K)")
    p2 <- mean_tile_plot(dist.df)
    p3 <- median_tile_plot(dist.df)
    p <- arrangeGrob(p1,p2,p3,ncol = 1)
    ggsave(plot=p,filename = plot.dir %&% "annot_counts_" %&% tiss %&% ".png",
           height = 10,width=6)
    ggsave(plot=p,filename = plot.dir %&% "annot_counts_" %&% tiss %&% ".pdf",
         height = 10,width=6)
  }
}

```



```{r}

save_composite_plots()


```
