==> tactical_analyses/03.2_annotation_plots.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Setup 

```{r}

"%&%" <- function(a,b) paste0(a,b)
library("tidyverse")
library("GenomicRanges")
library("viridis")
library("data.table")
library("plyr")

serv.dir <- "/home/jason/science/servers/FUSE5/"
proj.dir <- serv.dir %&% "projects/t2d_classification/"

work.dir <- proj.dir %&% "tactical_analyses/"
analysis.dir <- work.dir %&% "analysis_files/"
plot.dir <- work.dir %&% "plots/"


```


```{r}

annot.df <- fread(work.dir%&%
                    "analysis_files/annotation-scores-signals.txt")
names(annot.df)[1] <- "Locus.ID"
annot.df$strong_enhancer.Liver <- annot.df$`10_Active_enhancer_2.Liver` +
  annot.df$`9_Active_enhancer_1.Liver`
annot.df$strong_enhancer.Muscle <- annot.df$`10_Active_enhancer_2.Muscle` +
  annot.df$`9_Active_enhancer_1.Muscle`
annot.df$strong_enhancer.Adipose <- annot.df$`10_Active_enhancer_2.Adipose` +
  annot.df$`9_Active_enhancer_1.Adipose`
annot.df$strong_enhancer.Islet <- annot.df$`10_Active_enhancer_2.Islet` +
  annot.df$`9_Active_enhancer_1.Islet`


```


```{r}

cred.df <- fread(work.dir %&% "genetic_credible_sets/gencred.txt")
annot.df$maxppa <- map(annot.df$Locus.ID,function(id){
  filter(cred.df,CondID==id)$PPA %>% max(.)
}) %>% as.numeric(.)

class.df1 <- fread(analysis.dir%&%"classified-loci_weighted.txt")
class.df1$max.ppa <- map(class.df1$Locus.ID,function(id){
  sub <- filter(cred.df,CondID==id) %>% arrange(.,desc(PPA))
  sub$PPA[1]
  }) %>% as.numeric(.)
class.df2 <- fread(analysis.dir%&%"classified-loci_weighted_with-shared.txt")
class.df2$max.ppa <- map(class.df2$Locus.ID,function(id){
  sub <- filter(cred.df,CondID==id) %>% arrange(.,desc(PPA))
  sub$PPA[1]
  }) %>% as.numeric(.)

```


```{r}

simpleCap <- function(x) {
  s <- strsplit(x, " ")[[1]]
  paste(toupper(substring(s, 1,1)), substring(s, 2),
      sep="", collapse=" ")
}

```


# Sum plots 

```{r}

name.df <- data.frame(short=c("coding","1_Active_TSS","2_Weak_TSS",
                              "3_Flanking_TSS","5_Strong_transcription",
                              "6_Weak_transcription","8_Genic_enhancer",
                              "9_Active_enhancer_1","10_Active_enhancer_2",
                              "strong_enhancer",
                              "11_Weak_enhancer"),
                      long=c("Coding","Active TSS","Weak TSS",
                              "Flanking TSS","Strong transcription",
                              "Weak transcription","Genic enhancer",
                              "Strong enhancer (1)","Strong enhancer (2)",
                              "Strong enhancer (1,2)",
                              "Weak enhancer"),stringsAsFactors = F)

build_plot_df <- function(a.df){
  tiss.vec  <- c("islet","liver","muscle","adipose")
  plot.df <- c()
  for (tiss in tiss.vec){
    sub.df <- dplyr::select(a.df,contains(tiss))
    names(sub.df) <- map(names(sub.df),function(n){
      strsplit(n,".",fixed=TRUE)[[1]][1]
    }) %>% as.character(.)
    names(sub.df) <- map(names(sub.df),function(n){
      filter(name.df,short==n)$long 
    }) %>% as.character(.)
    mat <- as.matrix(sub.df)
    sums <- colSums(mat)
    build.df <- data.frame(Annotation=names(sub.df),Score.Sum=sums,Tissue=tiss,
                          stringsAsFactors = FALSE)  
    plot.df <- rbind(plot.df,build.df)
  }
  return(plot.df)
}


annot_plot <- function(plt.df){
  denom <- sum(plt.df$Score.Sum) - 
    (filter(plt.df,Annotation=="Strong enhancer (1,2)")$Score.Sum %>% sum(.))
  plt.df$Annotation <- factor(plt.df$Annotation,
                              levels=c("Coding","Active TSS","Weak TSS",
                              "Flanking TSS","Strong transcription",
                              "Weak transcription","Genic enhancer",
                              "Strong enhancer (1)","Strong enhancer (2)",
                              "Strong enhancer (1,2)",
                              "Weak enhancer"))
  p <- ggplot(data=plt.df,aes(x=Annotation,y=Score.Sum/denom)) + 
    geom_histogram(stat="identity",color="black",aes(fill=Tissue)) + 
    theme_bw() + 
    theme(axis.text.x=element_text(angle=30,vjust=0.6),
          legend.position = "none") + 
    scale_fill_manual(values=c("gold1","olivedrab2","brown","firebrick1")) + 
    ylab("Proportion of PPA in functional regions") + 
    coord_cartesian(ylim=c(0,0.5))
}


```



```{r}


library("gridExtra")
plt.df1 <- build_plot_df(annot.df)
plt.df2 <- build_plot_df(filter(annot.df,maxppa>0.5))
plt.df3 <- build_plot_df(filter(annot.df,maxppa>0.9))

plt1 <- annot_plot(plt.df1)
plt2 <- annot_plot(plt.df2)
plt3 <- annot_plot(plt.df3)


plt <- grid.arrange(plt1,plt2,plt3,nrow=1)
ggsave(plot=plt,filename=plot.dir%&%"annot_profile.png",
       width=16,height=5)
ggsave(plot=plt,filename=plot.dir%&%"annot_profile.pdf",
       width=16,height=5)

```



```{r}

prepare_toa_collapsed <- function(c.df){
  tiss.vec <- c("islet","liver","adipose","muscle","unclassified")
  score.sum <- c(c.df$islet %>% sum(.), 
                 c.df$liver %>% sum(.), 
                 c.df$adipose %>% sum(.),
                 c.df$muscle %>% sum(.),
                 c.df$unclassified %>% sum(.))
  score.prop <- score.sum / sum(score.sum)
  add.df <- data.frame(tissue=tiss.vec,score.sum,score.prop,stringsAsFactors = F)  
}

toa_collap_plot <- function(p.df){
  p.df$tissue <- map(p.df$tissue,function(x) {
    s <- strsplit(x, " ")[[1]]
    paste(toupper(substring(s, 1,1)), substring(s, 2),
        sep="", collapse=" ")
    }) %>% as.character(.)
  names(p.df)[1] <- "Tissue"
  plt <- ggplot(data=p.df,aes(x=Tissue,y=score.prop)) + 
    geom_histogram(stat="identity",color="black",aes(fill=Tissue),
                 binwidth=1,size=1) + 
    theme_bw()  + 
    theme(axis.text.x=element_text(angle=30,vjust=0.6)) + 
    scale_fill_manual(values=c("gold1","olivedrab2","brown","firebrick1","grey")) +
  ylab("Proportion of Total PPA") + 
  coord_cartesian(ylim=c(0,0.4))
}


```


```{r}

plt.df1b <- prepare_toa_collapsed(class.df1)
plt.df2b <- prepare_toa_collapsed(filter(class.df1,max.ppa>0.5))
plt.df3b <- prepare_toa_collapsed(filter(class.df1,max.ppa>0.9))

plt1b <- toa_collap_plot(plt.df1b)
ggsave(plot=plt1b,filename=plot.dir%&%"toa_profile_all.png",
       width=5,height=4)
ggsave(plot=plt1b,filename=plot.dir%&%"toa_profile_all.pdf",
       width=5,height=4)
plt2b <- toa_collap_plot(plt.df2b)
ggsave(plot=plt2b,filename=plot.dir%&%"toa_profile_50.png",
       width=5,height=4)
ggsave(plot=plt2b,filename=plot.dir%&%"toa_profile_50.pdf",
       width=5,height=4)
plt3b <- toa_collap_plot(plt.df3b)
ggsave(plot=plt3b,filename=plot.dir%&%"toa_profile_90.png",
       width=5,height=4)
ggsave(plot=plt3b,filename=plot.dir%&%"toa_profile_90.pdf",
       width=5,height=4)

```


# Summary values for manuscript 

```{r}

ids <- unique(cred.df$CondID)
snp.num <- map(ids,function(id){
  dim(filter(cred.df,CondID==id))[1] 
}) %>% as.integer(.)
summary(snp.num)
max.ppa <- map(ids,function(id){
  (filter(cred.df,CondID==id) %>% arrange(.,desc(PPA)))$PPA[1]
}) %>% as.numeric(.)
summary(max.ppa)

```


                              
```{r}

plt.df1b %>% arrange(.,desc(score.prop)) # islet 29%, unclass 20%

denom <- sum(plt.df1$Score.Sum) - 
    (filter(plt.df1,Annotation=="Strong enhancer (1,2)")$Score.Sum %>% sum(.))
plt.df1$prop <- plt.df1$Score.Sum / denom
plt.df1 %>% arrange(.,desc(prop)) # islet WT 21%, islet SE 10%
# WT total 
filter(plt.df1,Annotation=="Weak transcription")$prop %>% sum(.) # 50.7% 
filter(plt.df1,Annotation=="Coding")$prop %>% sum(.) # 0.02954429; 3.0% 
filter(plt.df1,Annotation=="Strong enhancer (1,2)")$prop %>% sum(.) # 17.6%
filter(plt.df1,Annotation=="Strong enhancer (1,2)",Tissue=="islet")$prop %>% sum(.) # 10.1%

```


```{r}

plt.df2b %>% arrange(.,desc(score.prop)) # islet 34%, unclass 14%, 11-21%

denom <- sum(plt.df2$Score.Sum) - 
    (filter(plt.df2,Annotation=="Strong enhancer (1,2)")$Score.Sum %>% sum(.))
plt.df2$prop <- plt.df2$Score.Sum / denom

plt.df2 %>% arrange(.,desc(prop)) # WT islet 17%, SE 15%
# WT total 
filter(plt.df2,Annotation=="Weak transcription")$prop %>% sum(.) # 40.4%
filter(plt.df2,Annotation=="Strong enhancer (1,2)")$prop %>% sum(.) # 24%
filter(plt.df2,Annotation=="Strong enhancer (1,2)",Tissue=="islet")$prop %>% sum(.) # 15%

```


```{r}

plt.df3b %>% arrange(.,desc(score.prop)) # islet 38%, unclass 5%, 14-23%

denom <- sum(plt.df3$Score.Sum) - 
    (filter(plt.df3,Annotation=="Strong enhancer (1,2)")$Score.Sum %>% sum(.))
plt.df3$prop <- plt.df3$Score.Sum / denom
plt.df3 %>% arrange(.,desc(prop)) # WT islet 12%, SE 16%
# WT total 
filter(plt.df3,Annotation=="Weak transcription")$prop %>% sum(.) # 31%
filter(plt.df3,Annotation=="Strong enhancer (1,2)")$prop %>% sum(.) # 26%
filter(plt.df3,Annotation=="Strong enhancer (1,2)",Tissue=="islet")$prop %>% sum(.) # 16%


```



```{r}
# islet to other tissue (next highest) factor
plt.df1b %>% arrange(.,desc(score.prop)) # islet = 0.2914650; liver =  0.1898216; factor = 1.54
plt.df2b %>% arrange(.,desc(score.prop)) # islet = 0.3413803; liver =  0.2139818; factor = 1.59
plt.df3b %>% arrange(.,desc(score.prop)) # islet = 0.37988524; liver =  0.23992068; factor = 1.58

#increase of 

```


# Strong Enhancer Analysis 

```{r}

isl.all <- filter(plt.df1,Annotation=="Strong enhancer (1,2)",Tissue=="islet")$prop %>% sum(.) # 10.06%
isl.50 <- filter(plt.df2,Annotation=="Strong enhancer (1,2)",Tissue=="islet")$prop %>% sum(.) # 15.0%
isl.90 <- filter(plt.df3,Annotation=="Strong enhancer (1,2)",Tissue=="islet")$prop %>% sum(.) # 16.3%

adi.all <- filter(plt.df1,Annotation=="Strong enhancer (1,2)",Tissue=="adipose")$prop %>% sum(.) # 4.2%
adi.50 <- filter(plt.df2,Annotation=="Strong enhancer (1,2)",Tissue=="adipose")$prop %>% sum(.) # 4.4%
adi.90 <- filter(plt.df3,Annotation=="Strong enhancer (1,2)",Tissue=="adipose")$prop %>% sum(.) # 4.8%

mus.all <- filter(plt.df1,Annotation=="Strong enhancer (1,2)",Tissue=="muscle")$prop %>% sum(.) # 0.8 %
mus.50 <- filter(plt.df2,Annotation=="Strong enhancer (1,2)",Tissue=="muscle")$prop %>% sum(.) # 1.2%
mus.90 <- filter(plt.df3,Annotation=="Strong enhancer (1,2)",Tissue=="muscle")$prop %>% sum(.) # 1.5%

liv.all <- filter(plt.df1,Annotation=="Strong enhancer (1,2)",Tissue=="liver")$prop %>% sum(.) # 2.5%
liv.50 <- filter(plt.df2,Annotation=="Strong enhancer (1,2)",Tissue=="liver")$prop %>% sum(.) # 3.4%
liv.90 <- filter(plt.df3,Annotation=="Strong enhancer (1,2)",Tissue=="liver")$prop %>% sum(.) # 3.8%

```


```{r}

tot.all <- isl.all + mus.all + adi.all  + liv.all
tot.50 <- isl.50 + mus.50 + adi.50  + liv.50
tot.90 <- isl.90 + mus.90 + adi.90  + liv.90

all.vec <- c(isl.all,liv.all,adi.all,mus.all)
all.vec / tot.all # islet = 57%; liver = 14%; adipose <- 24%; muscle <- 5% 

fifty.vec <- c(isl.50,liv.50,adi.50,mus.50)
fifty.vec / tot.50 # islet = 63%; liver = 14%; adipose <- 18%; muscle <- 5% 

ninety.vec <- c(isl.90,liv.90,adi.90,mus.90)
ninety.vec / tot.90 # islet = 62%; liver = 14%; adipose <- 18%; muscle <- 6% 

```



# Adjusting annotation proportions by genomic abundance proporition, within each tissue SEPARATELY!!

```{r}

count.df <- fread(analysis.dir %&% "annot_count.txt")

```


```{r}

append_count_proportions <- function(df,count.df){
  # df is plot.df for annotations 
  out.df <- c()
  for (i in 1:dim(df)[1]){
    row.df <- df[i,]
    annot <- row.df$Annotation; tiss <- row.df$Tissue
    t <- ifelse(tiss=="islet","Islet",
                ifelse(tiss=="muscle","Muscle",
                       ifelse(tiss=="adipose","Adipose",
                              ifelse(tiss=="liver","Liver",NA))))
    a <- ifelse(annot=="Strong enhancer (2)","Active_enhancer_2",
         ifelse(annot=="Weak enhancer","Weak_enhancer",
         ifelse(annot=="Active TSS","Active_TSS",
         ifelse(annot=="Weak TSS","Weak_TSS",
         ifelse(annot=="Flanking TSS","Flanking_TSS",
         ifelse(annot=="Strong transcription","Strong_transcription",
         ifelse(annot=="Weak transcription","Weak_transcription",
         ifelse(annot=="Genic enhancer","Genic_enhancer",
         ifelse(annot=="Strong enhancer (1)","Active_enhancer_1",
         ifelse(annot=="Coding","Coding",
         ifelse(annot=="Strong enhancer (1,2)",c("Active_enhancer_1","Active_enhancer_2"))))))))))))
    if (annot=="Strong enhancer (1,2)"){
      a <- c("Active_enhancer_1","Active_enhancer_2")
    }
    genome.prop <- ifelse(a=="Coding",0.01,(filter(count.df,Tissue==t,Annotation%in%a)$Proportion)) %>% sum(.)
    row.df$genome.prop <- genome.prop
    denom <- filter(df,Tissue==tiss,Annotation!="Strong enhancer (1,2)")$Score.Sum %>% 
      sum(.)
    adj.prop <- row.df$Score.Sum / denom
    row.df$adj.prop <- adj.prop
    out.df <- rbind(out.df,row.df)
  }
  out.df$ratio <- out.df$adj.prop / out.df$genome.prop
  return(out.df)
}

```



```{r}

library("ggrepel")

prop_plot <- function(prop.df){
  p <- ggplot(data=filter(prop.df,Annotation!="Coding"),
             aes(x=genome.prop,y=adj.prop,label=)) + geom_point() + 
    geom_abline(slope=1)  + 
    geom_text_repel(aes(label=Annotation),size=2)+
    facet_wrap(~Tissue,nrow=1) + 
    theme_classic() + 
    xlab("Within-tissue proportion of genomic regions") + 
    ylab("Within-tissue proportion\n of cumulative PPA")
  return(p)
}

```


```{r}

prop.df1 <- append_count_proportions(plt.df1,count.df)
prop.df2 <- append_count_proportions(plt.df2,count.df)
prop.df3 <- append_count_proportions(plt.df3,count.df)

pp1 <- prop_plot(prop.df1)
pp2<- prop_plot(prop.df2)
pp3 <- prop_plot(prop.df3)

```


```{r}

write.table(x=prop.df1,file=analysis.dir%&%"genome_adjusted_annotationPPA_allsigs.txt",sep="\t",quote=F,row.names=F)
write.table(x=prop.df2,file=analysis.dir%&%"genome_adjusted_annotationPPA_ppa50.txt",sep="\t",quote=F,row.names=F)
write.table(x=prop.df3,file=analysis.dir%&%"genome_adjusted_annotationPPA_ppa90.txt",sep="\t",quote=F,row.names=F)

```


```{r}

filter(prop.df1,Annotation=="Strong enhancer (1,2)")
filter(prop.df2,Annotation=="Strong enhancer (1,2)")
filter(prop.df3,Annotation=="Strong enhancer (1,2)")

```



```{r}

p <- grid.arrange(pp1,pp2,pp3,nrow=3)
ggsave(plot=p,filename=plot.dir%&%"genome_adjusted_annotationPPA.png",width=10,height = 8)
ggsave(plot=p,filename=plot.dir%&%"genome_adjusted_annotationPPA.pdf",width=10,height = 8)


```
