==> tactical_analyses/06.2_plot-coexpression.Rmd <==
---
title: "03.2b_plot-gene-coexpression.Rmd"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Setup 

```{r}

"%&%" <- function(a,b) paste0(a,b)
library("tidyverse")
library("GenomicRanges")
library("viridis")
library("data.table")
library("plyr")

serv.dir <- "/home/jason/science/servers/FUSE5/"
proj.dir <- serv.dir %&% "projects/t2d_classification/"

work.dir <- proj.dir %&% "tactical_analyses/"
analysis.dir <- work.dir %&% "analysis_files/"
plot.dir <- work.dir %&% "plots/"

count.df1 <- fread(work.dir%&%"analysis_files/classifier_counts.txt")
count.df2 <- fread(work.dir%&%"analysis_files/classifier_counts_with-shared.txt")

```


```{r}

simpleCap <- function(x) {
  s <- strsplit(x, " ")[[1]]
  paste(toupper(substring(s, 1,1)), substring(s, 2),
      sep="", collapse=" ")
}

append_classifer_counts <- function(w.df, count.df){
  names(count.df) <- names(count.df) %>% tolower(.)  
  names(count.df)[1] <- names(count.df)[1] %>% simpleCap(.)
  w.df$classifer.count <- map(1:dim(w.df)[1],function(i){
    row.df <- w.df[i,]
    count <- filter(count.df,Threshold==row.df$threshold) %>% 
     dplyr::select(.,one_of(as.character(unique(row.df$geneset)))) %>% as.integer(.)
    if (row.df$geneset=="peripheral"){
     count <- filter(count.df,Threshold==row.df$threshold) %>% 
        dplyr::select(.,one_of("adipose","liver","muscle")) %>% sum(.)  
    }
  return(count)
}) %>% as.integer(.)  
  return(w.df)
}

```


```{r}

plot_enrich <- function(plot.df){
  plt <- ggplot(data=plot.df,aes(x=tissue,y=enrich.factor)) + 
    geom_point(size=1.5,aes(shape=geneset,fill=log_pval),
               alpha=0.7)  # size=classifer.count,

  plt <- plt + facet_wrap(~threshold,nrow=1)
  plt <- plt + 
    scale_shape_manual(name="Gene set",
      values=c(21,22,23,24,25,4)) +
    scale_y_continuous(limits=c(0.8,2),breaks=seq(0.8,2,0.2)) + 
    #scale_fill_continuous(guide = guide_legend(nrow=1)) + 
    scale_fill_viridis(name="-log10(p)",direction = -1) + # ,guide = guide_legend(nrow=1)
    ylab("Enrichment") + xlab("Tissue") + 
    theme_bw() + 
    theme(axis.text.y=element_text(size=6.5,angle=0,vjust=0.5),
          axis.text.x=element_text(size=6.5,angle=30,vjust=0.5),
          strip.text = element_text(size=10),
          axis.title=element_text(size=12),
          legend.key.size = unit(1,"line")) + 
    scale_size(range=c(1,3),limits=c(0,350)) 
  return(plt)
}


plot_enrich_alt2 <- function(plot.df){
  plt <- ggplot(data=plot.df,aes(x=tissue,y=enrich.factor)) + 
    geom_point(size=1.5,aes(shape=geneset,fill=log_pval),
               alpha=0.7)  # size=classifer.count,

  plt <- plt + facet_wrap(~threshold,nrow=1)
  plt <- plt + 
    scale_shape_manual(name="Gene set",
      values=c(21,22,23,24,25,4)) +
    scale_y_continuous(limits=c(0.8,2),breaks=seq(0.8,2,0.2)) + 
    #scale_fill_continuous(guide = guide_legend(nrow=1)) + 
    scale_fill_viridis(name="-log10(p)",direction = -1) + # ,guide = guide_legend(nrow=1)
    ylab("Enrichment") + xlab("Tissue") + 
    theme_bw() + 
    theme(axis.text.y=element_text(size=6.5,angle=0,vjust=0.5),
          strip.text = element_text(size=10),
          axis.title=element_text(size=12),
          legend.key.size = unit(1,"line")) + 
    scale_size(range=c(1,3),limits=c(0,350)) + 
    coord_flip() 
  return(plt)
}

```


# Revised Attempt 

```{r}

#plot.dfV1 <- fread(analysis.dir %&%
#                 "coexpress-enrich_weighted_p10Coding_with-shared.txt") %>% 
#  na.omit(.) %>% append_classifer_counts(.,count.df2)
#pltV1 <- plot_enrich_alt2(plot.dfV1)
#ggsave(plot=pltV1,filename = plot.dir %&% "coexpress-enrich_weighted-p10Coding_with-shared_V1.png",
#       width=12,height=5.25)


```

1st Nearest 

```{r}

plot.df1 <- fread(analysis.dir %&%
                 "coexpress-enrich_weighted_p10Coding_with-shared_1stNearest.txt") %>% 
  na.omit(.) %>% append_classifer_counts(.,count.df2)
plt1 <- plot_enrich_alt2(plot.df1)
ggsave(plot=plt1,filename = plot.dir %&% "coexpress-enrich_weighted-p10Coding_with-shared_1stNearest.png",
       width=12,height=5.25)
ggsave(plot=plt1,filename = plot.dir %&% "coexpress-enrich_weighted-p10Coding_with-shared_1stNearest.pdf",
       width=12,height=5.25)
```


2nd Nearest 

```{r}

plot.df2 <- fread(analysis.dir %&%
                 "coexpress-enrich_weighted_p10Coding_with-shared_2ndNearest.txt") %>% 
  na.omit(.) %>% append_classifer_counts(.,count.df2)
plt2 <- plot_enrich_alt2(plot.df2)
ggsave(plot=plt2,filename = plot.dir %&% "coexpress-enrich_weighted-p10Coding_with-shared_2ndNearest.png",
       width=12,height=5.25)
ggsave(plot=plt2,filename = plot.dir %&% "coexpress-enrich_weighted-p10Coding_with-shared_2ndNearest.pdf",
       width=12,height=5.25)
```

3rd Nearest 

```{r}

plot.df3 <- fread(analysis.dir %&%
                 "coexpress-enrich_weighted_p10Coding_with-shared_3rdNearest.txt") %>% 
  na.omit(.) %>% append_classifer_counts(.,count.df2)
plt3 <- plot_enrich_alt2(plot.df3)
ggsave(plot=plt3,filename = plot.dir %&% "coexpress-enrich_weighted-p10Coding_with-shared_3rdNearest.png",
       width=12,height=5.25)
ggsave(plot=plt3,filename = plot.dir %&% "coexpress-enrich_weighted-p10Coding_with-shared_3rdNearest.pdf",
       width=12,height=5.25)
```


```{r}

keep.tissues <- c("Islet","Muscle...Skeletal","Liver",
                  "Adipose...Subcutaneous")

sub.df1 <- filter(plot.df1,tissue %in% keep.tissues)
sub.df1$tissue <- gsub("Muscle...Skeletal","Muscle",sub.df1$tissue)
sub.df1$tissue <- gsub("Adipose...Subcutaneous","Adipose",sub.df1$tissue)
plt.sub1 <- plot_enrich(sub.df1) +
  theme(axis.text.y=element_text(size=8,angle=0,vjust=0),
        axis.text.x=element_text(size=8,angle=30))
ggsave(plot=plt.sub1,filename = plot.dir %&% "coexpress-enrich_weighted-noCoding-fourtiss_with-shared_1stNearest.png",width=10,height=2.5)
ggsave(plot=plt.sub1,filename = plot.dir %&% "coexpress-enrich_weighted-noCoding-fourtiss_with-shared_1stNearest.pdf",width=10,height=2.5)

sub.df2 <- filter(plot.df2,tissue %in% keep.tissues)
sub.df2$tissue <- gsub("Muscle...Skeletal","Muscle",sub.df2$tissue)
sub.df2$tissue <- gsub("Adipose...Subcutaneous","Adipose",sub.df2$tissue)
plt.sub2 <- plot_enrich_alt2(sub.df2)
ggsave(plot=plt.sub2,filename = plot.dir %&% "coexpress-enrich_weighted-noCoding-fourtiss_with-shared_2ndNearest.png",width=8.5,height=2)

sub.df3 <- filter(plot.df3,tissue %in% keep.tissues)
sub.df3$tissue <- gsub("Muscle...Skeletal","Muscle",sub.df3$tissue)
sub.df3$tissue <- gsub("Adipose...Subcutaneous","Adipose",sub.df3$tissue)
plt.sub3 <- plot_enrich_alt2(sub.df3)
ggsave(plot=plt.sub3,filename = plot.dir %&% "coexpress-enrich_weighted-noCoding-fourtiss_with-shared_3rdNearest.png",width=6,height=2)

```



```{r}

sub.df1$Type <- "1st Nearest"
sub.df2$Type <- "2nd Nearest"
sub.df3$Type <- "3rd Nearest"
sub.df <- rbind(sub.df1,sub.df2,sub.df3)

p <- plot_enrich_alt2(sub.df) + facet_wrap(~Type+threshold) + 
  theme(strip.text.x = element_text(size=6,
                                    margin = margin(0.08,0,0.08,0, "cm")),
        axis.text.x=element_text(size=6))
ggsave(plot=p,filename = plot.dir %&% "coexpress-enrich_weighted-noCoding-fourtiss_with-shared_full.png",width=9,height=3)

```



## Manually inspecting values 

```{r}

filter(plot.df1,geneset=="islet",threshold=="0.8") %>% arrange(.,desc(enrich.factor)) %>% View(.)


```
