==> tactical_analyses/08.2_phys-enrich-plots.Rmd <==
---
title: "visualize-enrichments.Rmd"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# Setup 

```{r}

"%&%" <- function(a,b) paste0(a,b)

library("tidyverse")
library("data.table")
library("viridis")

serv.dir <- "/home/jason/science/servers/FUSE5/"
proj.dir <- serv.dir %&% "projects/t2d_classification/"
work.dir <- proj.dir %&% "tactical_analyses/"
plot.dir <- work.dir %&% "plots/"

```



# Plotting functions  


```{r}

phys_enrich_plot <- function(enrich.df,myname,nc=4){
  plt <- ggplot(data=enrich.df,aes(x=tissue,y=enrichment)) + 
    geom_point(aes(size=count,fill=-log(p,base=10)),shape=21) + 
    facet_wrap(~physiology+threshold,ncol=nc) + 
    theme_bw() + 
    scale_fill_gradient2(low=viridis(n=50)[48],high=viridis(n=50)[10],
                         mid = viridis(n=50)[40],
                         midpoint=-log(0.05,base=10),
                         name=expression(-log10(p))) + 
    scale_size(range=c(0.5,2.5),limits=c(0,250)) + 
    #scale_x_continuous(breaks=c(0,0.2,0.5,0.8)) + 
    xlab("Classified Tissue") + ylab("Enrichment") + 
    ggtitle(myname) +
    theme(axis.text.x=element_text(vjust=0.6,angle=30),
          panel.grid.minor = element_blank(),
          panel.grid.major.x=element_blank(),
          #legend.key = element_rect(color="black",size=3),
          strip.background =element_rect(fill="ghostwhite"),
          strip.text = element_text(colour = magma(n=2)[1]))
}


phys_enrich_plot_v2 <- function(enrich.df,myname,nc=5){
  ggplot(data=enrich.df,aes(x=physiology,y=enrichment)) + 
    geom_point(aes(size=counts,fill=-log(p,base=10)),shape=21) +
    facet_wrap(~tissue,ncol=nc) + 
    theme_bw() + 
    scale_fill_gradient2(low=viridis(n=50)[48],high=viridis(n=50)[10],
                         mid = viridis(n=50)[40],
                         midpoint=-log(0.05,base=10),
                         name=expression(-log10(p))) + 
    xlab("Classified Tissue") + ylab("Enrichment") + 
    ggtitle(myname) +
    theme(axis.text.x=element_text(vjust=0.6,angle=30),
          panel.grid.minor = element_blank(),
          panel.grid.major.x=element_blank(),
          strip.background =element_rect(fill="ghostwhite"),
          strip.text = element_text(colour = magma(n=2)[1]))
}

mod_thresh <- function(enrich.df){
  enrich.df$threshold <- gsub("assigned_","0.",enrich.df$threshold)
  return(enrich.df)
}


```


# Mahajan approach 

```{r}


df2 <- fread(work.dir %&% "analysis_files/phys82-enrich_weighted_with-shared.txt") %>%
 mod_thresh(.)
df2$physiology <- gsub("_"," ",df2$physiology)
df2$threshold <- as.numeric(df2$threshold)

count.df2 <- fread(work.dir%&%"analysis_files/classifier_counts_with-shared.txt")
names(count.df2) <- c("Threshold","adipose","islet",
                      "liver","muscle","shared","unclassified")

df2$count <- map(1:dim(df2)[1],function(i){
  row.df <- df2[i,]
  count <- filter(count.df2,Threshold==row.df$threshold) %>% 
    dplyr::select(.,one_of(as.character(unique(row.df$tissue)))) %>% as.integer(.)
  if (row.df$tissue=="peripheral"){
    count <- filter(count.df2,Threshold==row.df$threshold) %>% 
      dplyr::select(.,one_of("adipose","liver","muscle")) %>% sum(.)  
  }
  return(count)
}) %>% as.integer(.)

plt2 <- phys_enrich_plot(df2,myname="Physiology enrichment for classifed loci (weighted)")

ggsave(plot=plt2,filename = plot.dir %&% "enrichment-phys_weighted-82_with-shared.png",width=12,height=8)
ggsave(plot=plt2,filename = plot.dir %&% "enrichment-phys_weighted-82_with-shared.pdf",width=12,height=8)

```



## Alternative method (differences in TOA scores between phys groups)

```{r}

phys.counts <- fread(work.dir %&% "analysis_files/phys94-ld-assigned.txt")$Physiology %>% 
  table(.) %>% as.data.frame(.)
names(phys.counts) <- c("phys","count")

```


```{r}


df3 <- fread(work.dir %&% "analysis_files/phys-enrich_weighted-V2.txt") %>% mod_thresh(.)
df3$physiology <- gsub("_"," ",df3$physiology)
df3$counts <- map(df3$physiology,function(p){
  filter(phys.counts,phys==p)$count
}) %>% as.integer(.)


plt3 <- phys_enrich_plot_v2(df3,myname="Physiology enrichment for classifed loci (weighted)")
ggsave(plot=plt3,filename = plot.dir %&% "enrichment-phys_weighted-82-V2.png",width=12,height=3.5)
ggsave(plot=plt3,filename = plot.dir %&% "enrichment-phys_weighted-82-V2.pdf",width=12,height=3.5)
```


# Udler approach 


```{r}


df4 <- fread(work.dir %&% "analysis_files/phys-enrich_weighted_with-shared_UDLER.txt") %>%
 mod_thresh(.)
df4$physiology <- gsub("_"," ",df4$physiology)
df4$threshold <- as.numeric(df4$threshold)

count.df4 <- fread(work.dir%&%"analysis_files/classifier_counts_with-shared.txt")
names(count.df4) <- c("Threshold","adipose","islet",
                      "liver","muscle","shared","unclassified")

df4$count <- map(1:dim(df4)[1],function(i){
  row.df <- df4[i,]
  count <- filter(count.df4,Threshold==row.df$threshold) %>% 
    dplyr::select(.,one_of(as.character(unique(row.df$tissue)))) %>% as.integer(.)
  if (row.df$tissue=="peripheral"){
    count <- filter(count.df4,Threshold==row.df$threshold) %>% 
      dplyr::select(.,one_of("adipose","liver","muscle")) %>% sum(.)  
  }
  return(count)
}) %>% as.integer(.)

plt4 <- phys_enrich_plot(df4,myname="Physiology enrichment for classifed loci (weighted)")

ggsave(plot=plt4,filename = plot.dir %&% "enrichment-phys_weighted_with-shared_UDLER.png",width=12,height=8)
ggsave(plot=plt4,filename = plot.dir %&% "enrichment-phys_weighted_with-shared_UDLER.pdf",width=12,height=8)

```


```{r}

phys.counts <- fread(work.dir %&% "analysis_files/physUDLER-ld-assigned.txt")$Physiology %>% 
  table(.) %>% as.data.frame(.)
names(phys.counts) <- c("phys","count")

df5 <- fread(work.dir %&% "analysis_files/phys-enrich_weighted-V2_UDLER.txt") %>% mod_thresh(.)
df5$physiology <- gsub("_"," ",df5$physiology)
df5$counts <- map(df5$physiology,function(p){
  filter(phys.counts,phys==p)$count
}) %>% as.integer(.)


plt5 <- phys_enrich_plot_v2(df5,myname="Physiology enrichment for classifed loci (weighted)")
ggsave(plot=plt5,filename = plot.dir %&% "enrichment-phys_weighted-UDLER-V2.png",width=12,height=3)
ggsave(plot=plt5,filename = plot.dir %&% "enrichment-phys_weighted-UDLER-V2.pdf",width=12,height=3)
```
